Turn the cursor into an object.
The field no longer holds a state with names of object types, nor it
manipulates its objects. The cursor has been turned into its own object
and the field only holds a reference to it. It knows its position and
color. The field can draw either one object or all of them.